Refactor file existence checks and infos
To enhance code reusability and improve startup loading, the checks for the existence of critical program files (e.g., common passwords, user passwords, previously found passwords) were refactored into a new separate method. Now, if any file doesn't exist, it will be created and an informational message will be printed to the console. Also, a warning is now printed if the user has not provided any passwords, to inform of possible program non-functionality.