Add background file operations support
Implemented a new file operations queue in the ArchiveCracker application to improve performance by handling I/O operations in the background. Refactored the 'AppendToCommonPasswordsFile' and 'SaveFoundPasswords' methods to be handled asynchronously by the 'FileOperationsWorker' method. Also, added a new enum 'OperationType' and class 'FileOperation' in the models to manage the operations in this queue.