Make Options class concrete instead of abstract
The Options class in the ArchiveCracker project has been made concrete from being abstract. This was done because the Options class is never instantiated. Adding the comment "Resharper disable once ClassNeverInstantiated.Global" makes it clear that this class is intended to be a global non-instantiable helper. This also allows for possible future extension and utilization.