Remove ParserExtensions file and refactor main method
The `ParserExtensions.cs` file and its asynchronous parsing methods were removed as they were creating unnecessary complexity in our code. Now, the arguments in the main method of `Program.cs` are parsed synchronously which improves readablility and debugability. Particularly, now it involves direct casting of the result and a clear separation of success and error cases. Also, any parsing errors will cause the application to exit immediately with status 1, making it clear that an error has occured.