Update ArchivePasswordPair properties to init set
Made ArchivePasswordPair properties 'File' and 'Password' to have init only setters. This change assures that once a value is set, it cannot be changed again. This is particularly useful within the ArchiveCracker context, since the archive file and password should not be allowed to change once they've been set. Additionally, a comment was added to Options.cs to disable a ReSharper warning about the property that could be made read only.