Add VariationsGenerator project to simulate passwords variations
This update introduces the VariationsGenerator project, which was added to the ArchiveCracker solution. This tool can generate various permutation and variations of acquired passwords. These variations can potentially unlock archives, documents, or other assets locked with variations of known passwords. The performance and efficiency was achieved through the use of async programming and the ConcurrentQueue for dealing with the fluctuations in password generation rates and file writing rates.