Add command line options and exception handling
Introduced command line options to specify input and output files. Added exception handling to continue processing when an individual password encounters a problem. This change provides more flexibility to the user and increases the robustness of the program. The CommandLineParser package was added to easily parse the command line options.