Add new password variation strategies and logging
This commit includes adding several new password variation strategies to the VariationsGenerator project, namely SimpleVariationsStrategy, ReversedStrategy, CapitalizeLettersStrategy, AppendPrependDatesStrategy, MirrorPasswordStrategy, DoublePasswordStrategy, AppendPrependCommonSequencesStrategy, LeetSpeakStrategy, AppendNumbersAndSymbolsStrategy, CharacterReplacementStrategy. These strategies were added to improve the variation generation process by offering a number of different password manipulation approaches, hence increasing the chance to guess the correct passwords.

Additionally, the commit introduces logging through the use of the Serilog package. This provides improved tracing and debugging capabilities, making it easier to follow the application process and identify potential issues or bottlenecks. The logging was added to the main program and to the strategies to log essential steps.

A known issue with this commit is the removal of PasswordVariations file, this is a breaking change and will be fixed in the next commits.