Optimize password checking with parallel execution
Modified the password checking methods in the PasswordService to utilize parallel execution, significantly increasing the speed of password checking. The max degree of parallelism is dynamically determined by the processor count. Also, switched the _commonPasswords list data type to a ReadOnlyCollection for a performance improvement and better protection against modification. These changes will increase the overall performance of the ArchiveCracker software with faster password cracking.