Add PasswordCollector project and functionality
A new project called PasswordCollector has been added. This project includes
a Program.cs file which contains the main functionality of the project which is reading
and writing passwords to a file. This project will parse command line arguments and use
them to process files, logging progress and any exceptions that occur while running.

It also added PackageReferences to 'CommandLineParser' and 'Serilog' for command line parsing
and logging purposes respectively.

The project is added to the solution file and an Options.cs file is created to define the
required program options.

The changes facilitate password processing with improved performance, error handling and logging.