Moved attempted password count logger to the beginning of password check
The logging for attempted password and remaining password counts were previously logged after a password check fails or succeeds, making it hard to keep track of attempts when the application is prematurely stopped. For better tracking, these logs have been moved to take place right before a password is checked, not after.