Optimize archive extraction check
Refactored the extraction check in Archive strategies to only extract the first non-directory entry, rather than attempting to extract all entries. The change is to improve efficiency by avoiding needless extraction of all entries when only a successful extraction of one non-directory file is necessary to validate the password. Also added a quick return clause if all entries in the archive are directories.