"Refactor archive and password services to improve efficiency
This commits refactor the ArchiveService's constructor. It removes the protectedArchive argument as it is now being instantiated. Also, it removes the method checkPasswordsAsync as it has been relocated to PasswordService.
In the PasswordService, passwords' general method is added for both common and user passwords. Moreover, a new TaskPool instance was added. TaskPool class which is also initialized in this commit, to limit the degree of parallelism in our tasks.
In ZipArchiveStrategy there's an efficiency update; instead of validating all entries, gets only the first non-directory.
The ArchiveCracker main method now receives _protectedArchives from LoadArchivesAsync instead of directly accessing the field.
This change encourages efficiency in password checking which ultimately results in better performance."