Limit parallel archive cracking attempts
An upper limit (HardMaxSimultaneousArchivesToTry) has been introduced to control the maximum parallel archive cracking attempts irrespective of the available task pool capacity. This is due to the potential overload of the system when the task pool capacity is too large. In the CalculateMaxParallelArchives method, instead of directly returning half of the capacity, it now checks if the half of the capacity exceeds the HardMaxSimultaneousArchivesToTry. If so, it would return the hard max limit, otherwise, it returns the originally calculated capacity.