Add tracking for failed password attempts
In this commit, we have added a new feature to track instances when we fail to find the password for a specific archive file.

This was done by introducing a new operational type 'NotFound' in FileOperation.cs. In addition, a new file path 'notFoundPasswordsPath' was added to options in Options.cs to store passwords that weren't found. Modifications in the PasswordService.cs and FileService.cs files made this functionality possible. Lastly, a new literal for 'NotFoundPasswordsFileName' was included in Program.cs to handle instances when the password is not found.

This new feature enhances the performance analysis of the application by keeping track of unsolved cases. It also allows for potential re-processing of these cases in the future.